PC user flow
