+++ lib/common-lib/src/services/cohortService.js
@@ -200,119 +200,3 @@ export const syncOfflinePayload = async (params = {}, header = {}) => {
     return handleResponseException(e)
   }
 }
-export const registerPC = async (params = {}, header = {}) => {
-  try {
-    let headers = {
-      ...header,
-      'Content-Type': 'application/json',
-      Authorization: 'Bearer ' + localStorage.getItem('token')
-    }
-    const result = await post(
-      `${process.env.REACT_APP_API_URL}/program-coordinator/register/program_coordinator`,
-      params,
-      {
-        headers: headers
-      }
-    )
-    if (result?.data) {
-      return result?.data
-    } else {
-      return {}
-    }
-  } catch (e) {
-    return handleResponseException(e)
-  }
-}
-export const Pclist = async (params = {}, header = {}) => {
-  try {
-    let headers = {
-      ...header,
-      'Content-Type': 'application/json',
-      Authorization: 'Bearer ' + localStorage.getItem('token')
-    }
-    const result = await post(
-      `${process.env.REACT_APP_API_URL}/program-coordinator`,
-      params,
-      {
-        headers: headers
-      }
-    )
-    if (result?.data) {
-      return result?.data
-    } else {
-      return {}
-    }
-  } catch (e) {
-    return handleResponseException(e)
-  }
-}
-
-export const pcDetails = async (params = {}, header = {}) => {
-  try {
-    let headers = {
-      ...header,
-      'Content-Type': 'application/json',
-      Authorization: 'Bearer ' + localStorage.getItem('token')
-    }
-    const result = await post(
-      `${process.env.REACT_APP_API_URL}/program-coordinator/${params?.id}`,
-      params,
-      {
-        headers: headers
-      }
-    )
-    if (result?.data) {
-      return result?.data
-    } else {
-      return {}
-    }
-  } catch (e) {
-    return handleResponseException(e)
-  }
-}
-export const PcAvailableFacilitator = async (params = {}, header = {}) => {
-  try {
-    let headers = {
-      ...header,
-      'Content-Type': 'application/json',
-      Authorization: 'Bearer ' + localStorage.getItem('token')
-    }
-    const result = await post(
-      `${process.env.REACT_APP_API_URL}/program-coordinator/availablefacilitator/${params?.id}`,
-      params,
-      {
-        headers: headers
-      }
-    )
-    if (result?.data) {
-      return result?.data
-    } else {
-      return {}
-    }
-  } catch (e) {
-    return handleResponseException(e)
-  }
-}
-export const AddRemovePrerak = async (params = {}, header = {}) => {
-  try {
-    let headers = {
-      ...header,
-      'Content-Type': 'application/json',
-      Authorization: 'Bearer ' + localStorage.getItem('token')
-    }
-    const result = await patch(
-      `${process.env.REACT_APP_API_URL}/program-coordinator/${params?.id}`,
-      params,
-      {
-        headers: headers
-      }
-    )
-    if (result?.data) {
-      return result?.data
-    } else {
-      return {}
-    }
-  } catch (e) {
-    return handleResponseException(e)
-  }
-}
